Blobs get TTL upload should be uploaded to Cloud.
Fix blobs with ttlUpdate upload to cloud bug.
Refactored test Classes in ReplicationTest for their use in other modules.
Fix metric name typo in HelixVcrClusterMetrics.
