+++ ambry-replication/src/test/java/com.github.ambry.replication/MockFindTokenFactory.java
@@ -33,45 +33,3 @@ public class MockFindTokenFactory implements FindTokenFactory {
   }
 }
 
-class MockFindToken implements FindToken {
-  int index;
-  long bytesRead;
-
-  public MockFindToken(int index, long bytesRead) {
-    this.index = index;
-    this.bytesRead = bytesRead;
-  }
-
-  public MockFindToken(DataInputStream stream) throws IOException {
-    this.index = stream.readInt();
-    this.bytesRead = stream.readLong();
-  }
-
-  @Override
-  public byte[] toBytes() {
-    ByteBuffer byteBuffer = ByteBuffer.allocate(12);
-    byteBuffer.putInt(index);
-    byteBuffer.putLong(bytesRead);
-    return byteBuffer.array();
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    MockFindToken that = (MockFindToken) o;
-    return index == that.index && bytesRead == that.bytesRead;
-  }
-
-  public int getIndex() {
-    return index;
-  }
-
-  public long getBytesRead() {
-    return this.bytesRead;
-  }
-}
